Refactor forms into a module
