+++ cerberus/forms/booking.py
@@ -2,77 +2,9 @@
 from django import forms
 
 # Locals
-from .models import Booking, Charge, Contact, Customer, Invoice, Pet, Service, Vet
-from .utils import minimize_whitespace
-from .widgets import CheckboxDataOptionAttr, CheckboxTable, SelectDataAttrField, SingleMoneyWidget, TagsWidget
-
-
-class CustomerForm(forms.ModelForm):
-    class Meta:
-        model = Customer
-        fields = [
-            "first_name",
-            "last_name",
-            "other_names",
-            "invoice_address",
-            "invoice_email",
-            "active",
-            "vet",
-            "tags",
-        ]
-        widgets = {"tags": TagsWidget()}
-        help_texts = {"tags": None}
-
-
-class ContactForm(forms.ModelForm):
-    class Meta:
-        model = Contact
-        fields = [
-            "name",
-            "details",
-        ]
-        widgets = {
-            "details": forms.TextInput(
-                attrs={
-                    "x-data": "",
-                    "x-mask:dynamic": "$input.startsWith('0') ? '99999 999 999' : false",
-                }
-            ),
-        }
-
-
-class PetForm(forms.ModelForm):
-    class Meta:
-        model = Pet
-        fields = [
-            "name",
-            "dob",
-            "description",
-            "sex",
-            "neutered",
-            "social_media_concent",
-            "vet",
-            "treatment_limit",
-            "medical_conditions",
-            "allergies",
-            "tags",
-        ]
-        widgets = {"tags": TagsWidget()}
-        help_texts = {"tags": None}
-
-
-class VetForm(forms.ModelForm):
-    class Meta:
-        model = Vet
-        fields = [
-            "name",
-            "phone",
-            "details",
-        ]
-
-        widgets = {
-            "phone": forms.TextInput(attrs={"x-mask": "99999 999999", "x-data": ""}),
-        }
+from ..models import Booking
+from ..utils import minimize_whitespace
+from ..widgets import CheckboxDataOptionAttr, SelectDataAttrField, SingleMoneyWidget
 
 
 class BookingForm(forms.ModelForm):
@@ -200,84 +132,3 @@ class BookingForm(forms.ModelForm):
                 }
             ),
         }
-
-
-class InvoiceForm(forms.ModelForm):
-    class Meta:
-        model = Invoice
-        fields = [
-            "details",
-            "due",
-            "adjustment",
-            "customer_name",
-            "sent_to",
-            "invoice_address",
-            "send_notes",
-        ]
-        widgets = {
-            "adjustment": SingleMoneyWidget(),
-            "due": forms.DateInput(attrs={"type": "date"}),
-        }
-
-
-class ChargeForm(forms.ModelForm):
-    class Meta:
-        model = Charge
-        fields = ["name", "line", "quantity"]
-        widgets = {
-            "line": SingleMoneyWidget(attrs={"x-model.fill": "line", "min": "0"}),
-            "quantity": forms.NumberInput(attrs={"x-model.number.fill": "quantity"}),
-        }
-
-
-class ServiceForm(forms.ModelForm):
-    class Meta:
-        model = Service
-        fields = [
-            "name",
-            "length",
-            "booked_length",
-            "cost",
-            "cost_per_additional",
-            "max_pet",
-            "max_customer",
-            "display_colour",
-        ]
-        widgets = {
-            "cost": SingleMoneyWidget(),
-            "cost_per_additional": SingleMoneyWidget(),
-            "display_colour": forms.TextInput(attrs={"type": "color"}),
-        }
-
-
-class UninvoicedChargesForm(forms.Form):
-    customer = forms.ModelChoiceField(queryset=Customer.objects.all(), widget=forms.HiddenInput, required=True)
-    charges = forms.ModelMultipleChoiceField(
-        queryset=Charge.objects.none(),
-        widget=CheckboxTable(["name", "amount", "booking.date"]),
-    )
-
-    def __init__(self, *args, **kwargs):
-        customer = kwargs.pop("customer", None)
-        super().__init__(*args, **kwargs)
-
-        if self.is_bound and (customer_id := self.data.get("customer", None)):
-            self.set_customer(customer_id)
-        elif isinstance(customer, Customer):
-            self.set_customer(customer.pk)
-        elif isinstance(customer, int):
-            self.set_customer(customer)
-        else:
-            raise ValueError("No customer provided")
-
-    def set_customer(self, customer_id: int | None) -> None:
-        self.fields["customer"].initial = customer_id
-        self.fields["charges"].queryset = Charge.objects.filter(customer_id=customer_id, invoice__isnull=True)
-
-
-class InvoiceSendForm(forms.Form):
-    attributes = {"x-data": minimize_whitespace("""{ send: true }""")}
-
-    send_email = forms.BooleanField(required=False, widget=forms.CheckboxInput(attrs={"x-model.boolean.fill": "send"}))
-    to = forms.EmailField(required=True, widget=forms.EmailInput(attrs={":class": "{ 'display-none': ! send }"}))
-    send_notes = forms.CharField(required=False, widget=forms.Textarea(attrs={":class": "{ 'display-none': ! send }"}))
