+++ cerberus/forms/service.py
@@ -1,0 +1,26 @@
+# Django
+from django import forms
+
+# Locals
+from ..models import Service
+from ..widgets import SingleMoneyWidget
+
+
+class ServiceForm(forms.ModelForm):
+    class Meta:
+        model = Service
+        fields = [
+            "name",
+            "length",
+            "booked_length",
+            "cost",
+            "cost_per_additional",
+            "max_pet",
+            "max_customer",
+            "display_colour",
+        ]
+        widgets = {
+            "cost": SingleMoneyWidget(),
+            "cost_per_additional": SingleMoneyWidget(),
+            "display_colour": forms.TextInput(attrs={"type": "color"}),
+        }
